api/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==1.2.0
pydantic==2.6.1